refactor: update method caching and typing in core classes
The "memorizeMethods" function in the Cache class is updated to only cache methods if the configuration option "cached" is active. Also, The core classes Cache and Container have their typing for context map keys updated from String to string. A small refactoring was applied in the method extraction logic to pull out the config methods into a variable for clarity and readability.