feat: implement exclusion filter mode for caching system
A new feature has been added to the caching system which allows for an "exclude mode". In this mode, all methods, except specified ones, will be cached. The cached system tests were updated accordingly to ensure the proper functioning of this new feature. The exclude mode option was also added to the IServiceConfig interface. Minor documentation changes were made in the Container file.